Fix stack pointer size and more
Specifically, SP is now 8 bit instead of 16. StackPush and StackPull are fixed

+++ src/llvmes/dynarec/compiler.cpp
@@ -64,7 +64,7 @@ Compiler::Compiler(AST&& ast, const std::string& program_name)
     c->builder.SetInsertPoint(entry);
 
     // Initiate all variables to 0
-    c->reg_sp = c->builder.CreateAlloca(int16, 0, "SP");
+    c->reg_sp = c->builder.CreateAlloca(int8, 0, "SP");
     c->reg_x = c->builder.CreateAlloca(int8, 0, "X");
     c->reg_y = c->builder.CreateAlloca(int8, 0, "Y");
     c->reg_a = c->builder.CreateAlloca(int8, 0, "A");
